requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
numpy>=1.24.0
nba_api>=1.4.0
hmmlearn>=0.3.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0
requests>=2.31.0
google-auth>=2.20.0
google-auth-oauthlib>=1.0.0
google-api-python-client>=2.90.0